+++ EtcsServer/Database/Entity/Train.cs
@@ -3,9 +3,9 @@
     public class Train
     {
         public int TrainId { get; set; }
-        public int LengthMeters { get; set; }
-        public int MaxSpeedMps { get; set; }
-        public int BrakeWeight { get; set; }
+        public double LengthMeters { get; set; }
+        public double MaxSpeedMps { get; set; }
+        public double BrakeWeight { get; set; }
         public List<Message> Messages { get; set; }
     }
 }
